Fix up tests for new thumbnail scenario

+++ geonode/base/tests.py
@@ -8,9 +8,10 @@ class ThumbnailTests(TestCase):
         self.rb = ResourceBase.objects.create()
 
     def tearDown(self):
-        t = self.rb.thumbnail_set.get()
-        if t:
-            t.delete()
+        if self.rb.thumbnail_set.exists():
+            t = self.rb.thumbnail_set.get()
+            if t:
+                t.delete()
 
     def test_initial_behavior(self):
         self.assertFalse(self.rb.has_thumbnail())
